LSC/backend/requirements.txt
<<<<<<< HEAD
asgiref==3.9.1
channels==4.3.1
dj-rest-auth==7.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
=======
altgraph==0.17.4
amqp==5.3.1
APScheduler==3.11.0
asgiref==3.9.1
billiard==4.2.1
celery==5.4.0
certifi==2025.8.3
channels==4.3.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
dj-rest-auth==7.0.1
Django==5.0.7
django-apscheduler==0.7.0
django-celery-beat==2.6.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-auth==0.9.5
django-timezone-field==7.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
kombu==5.5.4
packaging==25.0
pefile==2023.2.7
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pywin32==311
pywin32-ctypes==0.2.3
redis==5.0.6
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
>>>>>>> develop
